refactor: moved table types to the component folder
