+++ react-client/src/components/design-system/table/TableRowActions.tsx
@@ -1,5 +1,5 @@
 import { IconButton } from "~/components/design-system";
-import { RowAction, TableRowActionsProps } from "../types/TableTypes";
+import { RowAction, TableRowActionsProps } from "./TableTypes";
 
 export const TableRowActions = ({
   onSelect,
