+++ react-client/src/components/design-system/table/TableTypes.ts
